Move object rendering into a module
