+++ src/render/object.rs
@@ -1,0 +1,272 @@
+use crate::render::{
+    Render,
+    read_shaders,
+    COLOR_FORMAT, DEPTH_FORMAT,
+    GlobalContext,
+};
+use m3d::NUM_COLOR_IDS;
+
+use wgpu;
+
+use std::mem;
+
+
+const COLOR_TABLE: [[u8; 2]; NUM_COLOR_IDS as usize] = [
+    [0, 0],   // reserved
+    [128, 3], // body
+    [176, 4], // window
+    [224, 7], // wheel
+    [184, 4], // defence
+    [224, 3], // weapon
+    [224, 7], // tube
+    [128, 3], // body red
+    [144, 3], // body blue
+    [160, 3], // body yellow
+    [228, 4], // body gray
+    [112, 4], // yellow (charged)
+    [0, 2],   // material 0
+    [32, 2],  // material 1
+    [64, 4],  // material 2
+    [72, 3],  // material 3
+    [88, 3],  // material 4
+    [104, 4], // material 5
+    [112, 4], // material 6
+    [120, 4], // material 7
+    [184, 4], // black
+    [240, 3], // body green
+    [136, 4], // skyfarmer kenoboo
+    [128, 4], // skyfarmer pipetka
+    [224, 4], // rotten item
+];
+
+#[repr(C)]
+#[derive(Clone, Copy)]
+pub struct Vertex {
+    pub pos: [i8; 4],
+    pub color: u32,
+    pub normal: [i8; 4],
+}
+
+#[derive(Clone, Copy)]
+pub struct Locals {
+    pub matrix: [[f32; 4]; 4],
+}
+
+pub struct Context {
+    pub uniform_buf: wgpu::Buffer,
+    pub bind_group: wgpu::BindGroup,
+    pub pipeline_layout: wgpu::PipelineLayout,
+    pub pipeline: wgpu::RenderPipeline,
+}
+
+impl Context {
+    fn create_pipeline(
+        layout: &wgpu::PipelineLayout,
+        device: &wgpu::Device,
+    ) -> wgpu::RenderPipeline {
+        let shaders = read_shaders("object", &[], device)
+            .unwrap();
+        device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+            layout,
+            vertex_stage: wgpu::PipelineStageDescriptor {
+                module: &shaders.vs,
+                entry_point: "main",
+            },
+            fragment_stage: wgpu::PipelineStageDescriptor {
+                module: &shaders.fs,
+                entry_point: "main",
+            },
+            rasterization_state: wgpu::RasterizationStateDescriptor {
+                front_face: wgpu::FrontFace::Ccw,
+                // original was not drawn with rasterizer, used no culling
+                cull_mode: wgpu::CullMode::None,
+                depth_bias: 0,
+                depth_bias_slope_scale: 0.0,
+                depth_bias_clamp: 0.0,
+            },
+            primitive_topology: wgpu::PrimitiveTopology::TriangleList,
+            color_states: &[
+                wgpu::ColorStateDescriptor {
+                    format: COLOR_FORMAT,
+                    alpha: wgpu::BlendDescriptor::REPLACE,
+                    color: wgpu::BlendDescriptor::REPLACE,
+                    write_mask: wgpu::ColorWriteFlags::all(),
+                },
+            ],
+            depth_stencil_state: Some(wgpu::DepthStencilStateDescriptor {
+                format: DEPTH_FORMAT,
+                depth_write_enabled: true,
+                depth_compare: wgpu::CompareFunction::LessEqual,
+                stencil_front: wgpu::StencilStateFaceDescriptor::IGNORE,
+                stencil_back: wgpu::StencilStateFaceDescriptor::IGNORE,
+                stencil_read_mask: !0,
+                stencil_write_mask: !0,
+            }),
+            index_format: wgpu::IndexFormat::Uint16,
+            vertex_buffers: &[
+                wgpu::VertexBufferDescriptor {
+                    stride: mem::size_of::<Vertex>() as u32,
+                    step_mode: wgpu::InputStepMode::Vertex,
+                    attributes: &[
+                        wgpu::VertexAttributeDescriptor {
+                            offset: 0,
+                            format: wgpu::VertexFormat::Char4,
+                            attribute_index: 0,
+                        },
+                        wgpu::VertexAttributeDescriptor {
+                            offset: 4,
+                            format: wgpu::VertexFormat::Uint,
+                            attribute_index: 1,
+                        },
+                        wgpu::VertexAttributeDescriptor {
+                            offset: 8,
+                            format: wgpu::VertexFormat::Uchar4Norm,
+                            attribute_index: 2,
+                        },
+                    ],
+                },
+            ],
+            sample_count: 1,
+        })
+    }
+
+    fn create_color_table(
+        encoder: &mut wgpu::CommandEncoder,
+        device: &wgpu::Device
+    ) -> (wgpu::TextureView, wgpu::Sampler) {
+        let extent = wgpu::Extent3d {
+            width: NUM_COLOR_IDS as u32,
+            height: 1,
+            depth: 1,
+        };
+        let texture = device.create_texture(&wgpu::TextureDescriptor {
+            size: extent,
+            array_size: 1,
+            dimension: wgpu::TextureDimension::D1,
+            format: wgpu::TextureFormat::Rg8Uint,
+            usage: wgpu::TextureUsageFlags::SAMPLED | wgpu::TextureUsageFlags::TRANSFER_DST,
+        });
+
+        let staging = device
+            .create_buffer_mapped(NUM_COLOR_IDS as usize, wgpu::BufferUsageFlags::TRANSFER_SRC)
+            .fill_from_slice(&COLOR_TABLE);
+        encoder.copy_buffer_to_texture(
+            wgpu::BufferCopyView {
+                buffer: &staging,
+                offset: 0,
+                row_pitch: NUM_COLOR_IDS as u32 * 2,
+                image_height: 1,
+            },
+            wgpu::TextureCopyView {
+                texture: &texture,
+                level: 0,
+                slice: 0,
+                origin: wgpu::Origin3d {
+                    x: 0.0,
+                    y: 0.0,
+                    z: 0.0,
+                },
+            },
+            extent,
+        );
+
+        let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
+            r_address_mode: wgpu::AddressMode::ClampToEdge,
+            s_address_mode: wgpu::AddressMode::ClampToEdge,
+            t_address_mode: wgpu::AddressMode::ClampToEdge,
+            mag_filter: wgpu::FilterMode::Nearest,
+            min_filter: wgpu::FilterMode::Nearest,
+            mipmap_filter: wgpu::FilterMode::Nearest,
+            lod_min_clamp: 0.0,
+            lod_max_clamp: 0.0,
+            max_anisotropy: 0,
+            compare_function: wgpu::CompareFunction::Always,
+            border_color: wgpu::BorderColor::TransparentBlack,
+        });
+        (texture.create_default_view(), sampler)
+    }
+
+    pub fn new(
+        init_encoder: &mut wgpu::CommandEncoder,
+        device: &wgpu::Device,
+        palette: &[[u8; 4]],
+        global: &GlobalContext,
+    ) -> Self {
+        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            bindings: &[
+                wgpu::BindGroupLayoutBinding { // object locals
+                    binding: 0,
+                    visibility: wgpu::ShaderStageFlags::VERTEX,
+                    ty: wgpu::BindingType::UniformBuffer,
+                },
+                wgpu::BindGroupLayoutBinding { // color map
+                    binding: 1,
+                    visibility: wgpu::ShaderStageFlags::VERTEX,
+                    ty: wgpu::BindingType::SampledTexture,
+                },
+                wgpu::BindGroupLayoutBinding { // palette map
+                    binding: 2,
+                    visibility: wgpu::ShaderStageFlags::VERTEX,
+                    ty: wgpu::BindingType::SampledTexture,
+                },
+                wgpu::BindGroupLayoutBinding { // main sampler
+                    binding: 3,
+                    visibility: wgpu::ShaderStageFlags::VERTEX,
+                    ty: wgpu::BindingType::Sampler,
+                },
+            ],
+        });
+        let uniform_buf = device.create_buffer(&wgpu::BufferDescriptor {
+            size: mem::size_of::<Locals>() as u32,
+            usage: wgpu::BufferUsageFlags::UNIFORM | wgpu::BufferUsageFlags::TRANSFER_DST,
+        });
+        let palette_view = Render::create_palette(
+            init_encoder, palette, device
+        );
+        let (color_table_view, color_table_sampler) = Self::create_color_table(
+            init_encoder, device
+        );
+        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            layout: &bind_group_layout,
+            bindings: &[
+                wgpu::Binding {
+                    binding: 0,
+                    resource: wgpu::BindingResource::Buffer {
+                        buffer: &uniform_buf,
+                        range: 0 .. mem::size_of::<Locals>() as u32,
+                    },
+                },
+                wgpu::Binding {
+                    binding: 1,
+                    resource: wgpu::BindingResource::TextureView(&color_table_view),
+                },
+                wgpu::Binding {
+                    binding: 2,
+                    resource: wgpu::BindingResource::TextureView(&palette_view),
+                },
+                wgpu::Binding {
+                    binding: 3,
+                    resource: wgpu::BindingResource::Sampler(&color_table_sampler),
+                },
+            ],
+        });
+        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+            bind_group_layouts: &[
+                &global.bind_group_layout,
+                &bind_group_layout,
+            ],
+        });
+        let pipeline = Self::create_pipeline(&pipeline_layout, device);
+
+        Context {
+            uniform_buf,
+            bind_group,
+            pipeline_layout,
+            pipeline,
+        }
+    }
+
+    pub fn reload(&mut self, device: &wgpu::Device) {
+        self.pipeline = Self::create_pipeline(&self.pipeline_layout, device);
+    }
+}
